requirements.txt
beautifulsoup4==4.12.3
psycopg2==2.9.10
python-dotenv==1.0.1
Requests==2.32.3
